fix: Trailing whitespace highlgihts correctly
I have a theory that this was only working because it was interacting
with my vimrc. I needed to blast it away from that in order to know that
trailing whitespace matcher needed to be defined _after_ the highlight.

Before this, you'd get an error on launch. Since this, it justworks(tm)