refactor: :highlight format string constructed once for text
I'm trying to pivot colour schemes. To do this I need to name the parts
of my current one. This is going to take a while and I'm on a bus so I
don't want to lose my work.

TIL that vim's :map function will mutate the array we're mapping over,
but neovim will return a copy. Passing dictionaries around to construct
a colour scheme is looking a little brittle, so saving them out as
strings should be a bit easier to manage.